fix use 1 file for vertex success handler
